// 04-country-app/src/app/countries/interfaces/cache-store.interface.ts
import { Country } from "./country.interface";
import { Region } from "./region.type";

export interface CashStore {
    byCapital: TermCountries;
    byCountry: TermCountries;    
    byRegion: RegionCountries;
}

export interface TermCountries {   
    term: string;
    countries: Country[];    
}

export interface RegionCountries {   
    term: Region;
    countries: Country[];    
}